+++ frontend/src/components/Header.js
@@ -1,9 +1,21 @@
-import React from 'react';
-import logo from './Gather.png'; 
-import accountIcon from './account.png';
-import './Header.css';
+import React, { useEffect } from "react";
+import logo from "../images/Gather.png";
+import accountIcon from "../images/account.png";
+import "../css/Header.css";
 
 function Header() {
+  useEffect(() => {
+    const account = document.getElementById("account-icon");
+    const menuContainer = document.getElementById("menuContainerNotConnected");
+
+    account.addEventListener("mouseenter", () => {
+      menuContainer.style.transform = "translatex(0px)";
+    });
+    menuContainer.addEventListener("mouseleave", () => {
+      menuContainer.style.transform = "translatex(300px)";
+    });
+  }, []);
+
   return (
     <header>
       <nav>
@@ -11,12 +23,26 @@ function Header() {
           <img src={logo} alt="Logo" />
         </div>
         <ul>
-          <li><a href="/">Accueil</a></li>
+          <li>
+            <a href="/">Accueil</a>
+          </li>
         </ul>
-		<div className="account-icon">
+        <div className="account-icon" id="account-icon">
           <a target="_blank" href="login-or-sign-in">
             <img src={accountIcon} alt="Account" />
           </a>
+        </div>
+        <div
+          className="menuContainer not-connected"
+          id="menuContainerNotConnected"
+        >
+          <ul>
+            <li>
+              <a target="_blank" href="login-or-sign-in">
+                <span>Se Connecter</span>
+              </a>
+            </li>
+          </ul>
         </div>
       </nav>
     </header>
