+++ frontend/src/css/Container.css
@@ -1,4 +1,4 @@
-.dashContainer{
+.dashContainer {
   width: calc(100% - 100px);
   height: 100%;
   padding: 20px;
@@ -75,7 +75,7 @@
   box-shadow: 5px 5px 10px rgba(255, 0, 0, 0.6);
 }
 
-.topContainer .profileContainer .profileImage img{
+.topContainer .profileContainer .profileImage img {
   object-fit: cover;
   width: 100;
   height: 100%;
@@ -95,7 +95,7 @@
   transition: transform 0.5s ease;
 }
 
-.topContainer .profileContainer .menuContainer {
+/* .topContainer .profileContainer .menuContainer {
   position: absolute;
   right: 10px;
   top: 50px;
@@ -125,21 +125,21 @@
 
 .topContainer .profileContainer .menuContainer ul li:hover{
   background: palevioletred;
-  color: black;
+  color: black;   
   transition: 0.3s;
-}
+} */
 
-@media screen and (max-width: 525px){
-  .dashContainer{
-    width: calc(100% - 68px)
+@media screen and (max-width: 525px) {
+  .dashContainer {
+    width: calc(100% - 68px);
   }
-  .topContainer .inputBox{
-  width: 120px;
-  z-index: 101;
+  .topContainer .inputBox {
+    width: 120px;
+    z-index: 101;
   }
 
-  .topContainer .inputBox input:focus{
+  .topContainer .inputBox input:focus {
     width: 320px;
     transition: width 0.3s cubic-bezier(0.17, 0.81, 0.38, 1.18);
-    }
+  }
 }
