+++ frontend/src/css/Header.css
@@ -1,4 +1,10 @@
 /* header.css */
+@font-face {
+  font-family: "Michroma";
+  src: url("../fonts/Michroma-Regular.ttf") format("truetype");
+  /* Ajoutez d'autres variantes de police ici si nÃ©cessaire */
+}
+
 header {
   background-color: white;
   padding: 10px;
@@ -8,7 +14,7 @@ header {
 nav {
   display: flex;
   align-items: center;
-  max-height: 60px;
+  justify-content: space-between;
 }
 
 .logo img {
@@ -41,5 +47,67 @@ a:hover {
 
 .account-icon img {
   width: 70px; /* Ajustez la taille de l'icône du compte selon vos besoins */
+  transition: transform 0.5s ease;
+  cursor: pointer;
+  font-size: 12px;
+  margin: 0px 10px;
 }
 
+.menuContainer {
+  position: absolute;
+  margin-top: 25px;
+  padding: 10px;
+  right: 10px;
+  top: 50px;
+  background: blueviolet;
+  border-radius: 0.5em;
+  width: 200px;
+  box-shadow: 5px 5px 10px rgba(255, 0, 0, 0.6);
+  z-index: 1;
+  transform: translatex(300px);
+  transition: transform 0.3s cubic-bezier(0.17, 0.81, 0.38, 1.18);
+}
+
+.menuContainer ul {
+  width: 100%;
+  height: 100%;
+  display: flex;
+  flex-direction: column;
+  font-family: "Michroma", sans-serif;
+  font-size: 12px;
+}
+
+.menuContainer ul li {
+  list-style: none;
+  margin: 5px 0;
+  padding: 10px 20px;
+  cursor: pointer;
+  text-align: left;
+}
+
+.menuContainer ul li:hover {
+  background: palevioletred;
+  color: black;
+  transition: 0.3s;
+}
+
+.menuContainer ul li a {
+  display: block;
+  text-decoration: none;
+  color: inherit;
+  width: 100%;
+  height: 100%;
+}
+
+.menuContainer ul li a:hover {
+  text-decoration: underline;
+  color: inherit;
+}
+
+.menuContainer .connected {
+  background: blueviolet;
+}
+
+.menuContainer .not-connected {
+  background: blueviolet;
+}
